requirements.txt
pycrypto
flask
Flask-Cors